Fix warning about %lu being used with an int in 32 bit wxOSX build.
NSUInteger is unsigned long in 64 bit builds, but unsigned int in 32 bits, so
case it to long explicitly to avoid warnings about using it with %lu format
specifier.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_3_0_BRANCH@76284 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/osx/cocoa/dataview.mm
@@ -179,7 +179,8 @@ inline wxDataViewItem wxDataViewItemFromMaybeNilItem(id item)
     // not just some arbitrary object, so we serialize the pointer into the
     // string. Notice the use of NSInteger which is big enough to store a
     // pointer in both 32 and 64 bit builds.
-    return [NSString stringWithFormat:@"%lu", reinterpret_cast<NSUInteger>(column)];
+    return [NSString stringWithFormat:@"%lu",
+                (unsigned long)reinterpret_cast<NSUInteger>(column)];
 }
 
 -(id) initWithColumnPointer:(const wxDataViewColumn*)column
